Title: 최솟값 3개, Time: 995ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240723/최솟값 3개/top-3-smallest-number.java
@@ -9,16 +9,16 @@ public class Main {
         int n = Integer.parseInt(br.readLine());
 
         StringTokenizer st = new StringTokenizer(br.readLine());
-        PriorityQueue<Integer> pq = new PriorityQueue<>();
+        PriorityQueue<Long> pq = new PriorityQueue<>();
 
         for (int i = 0; i < n; i++) {
-            pq.add(Integer.parseInt(st.nextToken()));
+            pq.add(Long.parseLong(st.nextToken()));
             if (pq.size() < 3) {
                 System.out.println(-1);
             }else{
-                int val1 = pq.poll();
-                int val2 = pq.poll();
-                int val3 = pq.poll();
+                long val1 = pq.poll();
+                long val2 = pq.poll();
+                long val3 = pq.poll();
                 System.out.println(val1 * val2 * val3);
                 pq.add(val1);
                 pq.add(val2);
